Module dex_stable::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_liquidity

Supply amounts of currencies to the pool.

add_pool_and_base_pool_liquidity

Supply amounts of currencies to the pool which contains the lp currency of the base pool.

create_base_pool

Create a stable amm pool.

create_meta_pool

Create a stable amm meta pool.

ramp_a

Start ramping up or down A parameter towards given future_a and future_a_time

remove_liquidity

Remove liquidity from a pool.

remove_liquidity_imbalance

Remove liquidity from a pool to the specify amounts of currencies.

remove_liquidity_one_currency

Remove liquidity from a pool to get one currency.

remove_pool_and_base_pool_liquidity

Remove liquidity from a pool which contains the lp currency of the base pool.

remove_pool_and_base_pool_liquidity_one_currency

Remove liquidity from a pool which contains the lp currency of the base pool to get one currency.

set_admin_fee

Update admin fee of the pool.

set_swap_fee

Update fee of the pool.

stop_ramp_a

Stop ramping A parameter.

swap

Swap a amounts of currencies to get other.

swap_meta_pool_underlying

Warning: Doc-Only

swap_pool_from_base

Swap the currency from basic pool to get amounts of other currency in pool. to get one currency.

swap_pool_to_base

Swap the currency from pool to get amounts of other currency in basic pool. to get one currency.

update_fee_receiver

Update admin fee receiver of the pool.

withdraw_admin_fee

Withdraw the admin fee from pool to admin fee receiver.