Function dex_stable::pallet::dispatchables::remove_liquidity_one_currency

pub fn remove_liquidity_one_currency<T: Config>(
    pool_id: T::PoolId,
    lp_amount: u128,
    index: u32,
    min_amount: u128,
    to: T::AccountId,
    deadline: BlockNumberFor<T>
)

Remove liquidity from a pool to get one currency.

Argument

pool_id: The id of pool.

lp_amount: The amounts of lp currency.

index: The index of receive currency.

min_amount: The min amounts of received currency;

deadline: Height of the cutoff block of this transaction

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::remove_liquidity_one_currency.