Struct dex_stable::pallet::Pallet

pub struct Pallet<T>(_);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn get_virtual_price(pool_id: T::PoolId) -> u128

pub fn get_a(pool_id: T::PoolId) -> u128

pub fn get_a_precise_by_id(pool_id: T::PoolId) -> u128

pub fn get_currencies(pool_id: T::PoolId) -> Vec<T::CurrencyId>

pub fn get_currency_index( pool_id: T::PoolId, currency_id: T::CurrencyId ) -> Option<u32>

pub fn get_currency(pool_id: T::PoolId, index: u32) -> Option<T::CurrencyId>

pub fn get_lp_currency(pool_id: T::PoolId) -> Option<T::CurrencyId>

pub fn get_currency_precision_multipliers(pool_id: T::PoolId) -> Vec<u128>

pub fn get_currency_balances(pool_id: T::PoolId) -> Vec<u128>

pub fn get_number_of_currencies(pool_id: T::PoolId) -> u32

pub fn get_admin_balances(pool_id: T::PoolId) -> Vec<u128>

impl<T: Config> Pallet<T>

pub fn calculate_meta_swap_underlying( pool_id: T::PoolId, in_amount: u128, currency_index_from: usize, currency_index_to: usize ) -> Result<u128, DispatchError>

impl<T: Config> Pallet<T>

pub fn create_base_pool( origin: OriginFor<T>, currency_ids: Vec<T::CurrencyId>, currency_decimals: Vec<u32>, a: u128, fee: u128, admin_fee: u128, admin_fee_receiver: T::AccountId, lp_currency_symbol: Vec<u8> ) -> DispatchResult

Create a stable amm pool.

Only admin can create pool.

Argument

currency_ids: The currencies will be join the created pool.

currency_decimals: The currencies corresponding decimals.

lp_currency_id: The specify lp currency id of the created pool.

a: The initial A of created pool.

fee: The swap fee of created pool.

admin_fee: The admin fee of created pool.

admin_fee_receiver: The admin fee receiver of created pool.

lp_currency_symbol: The symbol of created pool lp currency.

lp_currency_decimal: The decimal of created pool lp currency.

pub fn create_meta_pool( origin: OriginFor<T>, currency_ids: Vec<T::CurrencyId>, currency_decimals: Vec<u32>, a: u128, fee: u128, admin_fee: u128, admin_fee_receiver: T::AccountId, lp_currency_symbol: Vec<u8> ) -> DispatchResult

Create a stable amm meta pool.

Only admin can create pool.

Argument

currency_ids: The currencies will be join the created pool.

currency_decimals: The currencies corresponding decimals.

lp_currency_id: The specify lp currency id of the created pool.

a: The initial A of created pool.

fee: The swap fee of created pool.

admin_fee: The admin fee of created pool.

admin_fee_receiver: The admin fee receiver of created pool.

lp_currency_symbol: The symbol of created pool lp currency.

lp_currency_decimal: The decimal of created pool lp currency.

pub fn add_liquidity( origin: OriginFor<T>, pool_id: T::PoolId, amounts: Vec<u128>, min_mint_amount: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Supply amounts of currencies to the pool.

Argument

pool_id: The id of pool.

amounts: Supply amounts of currencies.

min_mint_amount: The min amount of lp currency get.

deadline: Height of the cutoff block of this transaction

pub fn swap( origin: OriginFor<T>, pool_id: T::PoolId, from_index: u32, to_index: u32, in_amount: u128, min_out_amount: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Swap a amounts of currencies to get other.

Argument

pool_id: The id of pool.

from_index: The index of swap currency id.

to_index: The index of receive currency id.

in_amount: The amounts of currencies swap.

min_mint_amount: The min amount of receive currency.

deadline: Height of the cutoff block of this transaction

pub fn remove_liquidity( origin: OriginFor<T>, pool_id: T::PoolId, lp_amount: u128, min_amounts: Vec<u128>, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Remove liquidity from a pool.

Argument

pool_id: The id of pool.

lp_amount: The amounts of lp currency.

min_amounts: The min amounts of pool’s currencies to get.

deadline: Height of the cutoff block of this transaction

pub fn remove_liquidity_one_currency( origin: OriginFor<T>, pool_id: T::PoolId, lp_amount: u128, index: u32, min_amount: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Remove liquidity from a pool to get one currency.

Argument

pool_id: The id of pool.

lp_amount: The amounts of lp currency.

index: The index of receive currency.

min_amount: The min amounts of received currency;

deadline: Height of the cutoff block of this transaction

pub fn remove_liquidity_imbalance( origin: OriginFor<T>, pool_id: T::PoolId, amounts: Vec<u128>, max_burn_amount: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Remove liquidity from a pool to the specify amounts of currencies.

Argument

pool_id: The id of pool.

amounts: The specify amounts of receive currencies.

max_burn_amount: The max amount of burned lp currency.

deadline: Height of the cutoff block of this transaction

pub fn add_pool_and_base_pool_liquidity( origin: OriginFor<T>, pool_id: T::PoolId, base_pool_id: T::PoolId, meta_amounts: Vec<u128>, base_amounts: Vec<u128>, min_to_mint: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Supply amounts of currencies to the pool which contains the lp currency of the base pool.

Argument

pool_id: The id of pool.

base_pool_id: The id of base pool.

meta_amounts: Supply amounts of currencies to pool. The last element must be zero.

base_amounts: Supply amounts of currencies to base pool.

min_to_mint: The min amount of pool lp currency get.

deadline: Height of the cutoff block of this transaction.

pub fn remove_pool_and_base_pool_liquidity( origin: OriginFor<T>, pool_id: T::PoolId, base_pool_id: T::PoolId, amount: u128, min_amounts_meta: Vec<u128>, min_amounts_base: Vec<u128>, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Remove liquidity from a pool which contains the lp currency of the base pool.

Argument

pool_id: The id of pool.

base_pool_id: The id of base pool.

amount: The amounts of lp currency to burn.

min_amounts_meta: The min amounts of pool’s currencies to get.

min_amounts_base: The min amounts of basic pool’s currencies to get.

deadline: Height of the cutoff block of this transaction.

pub fn remove_pool_and_base_pool_liquidity_one_currency( origin: OriginFor<T>, pool_id: T::PoolId, base_pool_id: T::PoolId, amount: u128, i: u32, min_amount: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Remove liquidity from a pool which contains the lp currency of the base pool to get one currency.

Argument

pool_id: The id of pool.

base_pool_id: The id of base pool.

amount: The amounts of lp currency to burn.

i: The index of target currency in basic pool.

min_amount: The min amounts of received currency.

deadline: Height of the cutoff block of this transaction.

pub fn swap_pool_from_base( origin: OriginFor<T>, pool_id: T::PoolId, base_pool_id: T::PoolId, in_index: u32, out_index: u32, dx: u128, min_dy: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Swap the currency from basic pool to get amounts of other currency in pool. to get one currency.

Argument

pool_id: The id of pool.

base_pool_id: The id of base pool.

in_index: The index of swap currency in basic pool.

out_index: The index of target currency in pool.

dx: The amounts of swap currency.

min_dy: The min amounts of target currency.

deadline: Height of the cutoff block of this transaction.

pub fn swap_pool_to_base( origin: OriginFor<T>, pool_id: T::PoolId, base_pool_id: T::PoolId, in_index: u32, out_index: u32, dx: u128, min_dy: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

Swap the currency from pool to get amounts of other currency in basic pool. to get one currency.

Argument

pool_id: The id of pool.

base_pool_id: The id of base pool.

in_index: The index of swap currency in basic pool.

out_index: The index of target currency in pool.

dx: The amounts of swap currency.

min_dy: The min amounts of target currency.

deadline: Height of the cutoff block of this transaction.

pub fn swap_meta_pool_underlying( origin: OriginFor<T>, pool_id: T::PoolId, in_index: u32, out_index: u32, dx: u128, min_dy: u128, to: T::AccountId, deadline: BlockNumberFor<T> ) -> DispatchResult

pub fn update_fee_receiver( origin: OriginFor<T>, pool_id: T::PoolId, fee_receiver: <T::Lookup as StaticLookup>::Source ) -> DispatchResult

Update admin fee receiver of the pool.

Only called by admin.

Argument

pool_id: The id of pool.

fee_receiver: The new admin fee receiver of this pool.

pub fn set_swap_fee( origin: OriginFor<T>, pool_id: T::PoolId, new_swap_fee: u128 ) -> DispatchResult

Update fee of the pool.

Only called by admin.

Argument

pool_id: The id of pool.

new_swap_fee: The new swap fee of this pool.

pub fn set_admin_fee( origin: OriginFor<T>, pool_id: T::PoolId, new_admin_fee: u128 ) -> DispatchResult

Update admin fee of the pool.

Only called by admin.

Argument

pool_id: The id of pool.

new_admin_fee: The new admin fee of this pool.

pub fn ramp_a( origin: OriginFor<T>, pool_id: T::PoolId, future_a: u128, future_a_time: u128 ) -> DispatchResult

Start ramping up or down A parameter towards given future_a and future_a_time

Only called by admin. Checks if the change is too rapid, and commits the new A value only when it falls under the limit range.

Argument

pool_id: The id of pool.

future_a: The new A to ramp towards.

future_a_time: Timestamp when the new A should be reached

pub fn stop_ramp_a(origin: OriginFor<T>, pool_id: T::PoolId) -> DispatchResult

Stop ramping A parameter.

Only called by admin.

Argument

pool_id: The id of pool.

pub fn withdraw_admin_fee( origin: OriginFor<T>, pool_id: T::PoolId ) -> DispatchResult

Withdraw the admin fee from pool to admin fee receiver.

Can called by anyone.

Argument

pool_id: The id of pool.

impl<T: Config> Pallet<T>

pub fn next_pool_id() -> T::PoolId

An auto-generated getter for NextPoolId.

impl<T: Config> Pallet<T>

pub fn pools<KArg>( k: KArg ) -> Option<Pool<T::PoolId, T::CurrencyId, T::AccountId, T::PoolCurrencyLimit, T::PoolCurrencySymbolLimit>>where KArg: EncodeLike<T::PoolId>,

An auto-generated getter for Pools.

impl<T: Config> Pallet<T>

pub fn lp_currencies<KArg>(k: KArg) -> Option<T::PoolId>where KArg: EncodeLike<T::CurrencyId>,

An auto-generated getter for LpCurrencies.

impl<T: Config> Pallet<T>

pub fn pool_account_id(pool_id: &T::PoolId) -> T::AccountId

only use two byte prefix to support 16 byte account id (used by test) “modl” ++ “dex/stbl” is 12 bytes, and 4 bytes remaining for PoolId